requirements.txt
gradio>=4.44.0
torch>=2.1.0
diffusers>=0.21.4
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
Pillow>=10.0.0
spaces>=0.16.0
huggingface_hub>=0.21.0
peft>=0.5.0 